requirements.txt
clipboard==0.0.4
MouseInfo==0.1.3
Pillow==8.0.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dotenv==0.15.0
PyTweening==1.0.3
selenium==3.141.0
urllib3==1.26.2
